Implement options parameter for launch() function
Instead of passing magic integers to every function, constants
are defined to different functionalities. Also, defined constants
in the standard are also used.

These constants are passed through a single variable, so that they
can be or'd together. Inside the function, bitwise and helps
understanding which options were set.

Reference:
http://stackoverflow.com/a/8778056/3578554

Also, the `>&` redirection (stdout & stderr) redirection operator
was implemented, using the above philosophy.