khol.c: Fix command strip on history recollection.
Previously, when the user would execute a large command like
`ls -s -w 1`, and then `!-1` followed by `!-2`, this would cause
command to get stripped to only to the first token, `ls`.

This happened because history_get() would return a pointer, which
when split, would cause the actual history to mutate.

The fix was pretty simple, to copy the string first to a temp
location, and then split it.